ai/requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.13
langchain-postgres>=0.0.2
python-dotenv>=1.0.0
openai>=1.10.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
langchain-experimental>=0.0.37
langgraph>=0.0.19